safari-534.53-branch/LayoutTests/storage/indexeddb/mozilla/create-index-null-name-expected.txt

Test IndexedDB's creating index with null name

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('a', { keyPath: 'id', autoIncrement: true });
Expecting exception from index = objectStore.createIndex(null, null);
PASS Exception was thrown.
PASS code is IDBDatabaseException.NON_TRANSIENT_ERR
PASS successfullyParsed is true

TEST COMPLETE
